% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsestr-package.R
\docType{package}
\name{palimpsestr-package}
\alias{palimpsestr}
\alias{palimpsestr-package}
\title{palimpsestr: Probabilistic Decomposition of Archaeological Palimpsests}
\description{
Probabilistic framework for the analysis of archaeological palimpsests
based on the Stratigraphic Entanglement Field (SEF).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/enzococca/palimpsestr}
  \item Report bugs at \url{https://github.com/enzococca/palimpsestr/issues}
}

}
\author{
\strong{Maintainer}: Enzo Cocca \email{enzo.ccc@gmail.com} (\href{https://orcid.org/0000-0002-8096-2810}{ORCID})

Authors:
\itemize{
  \item Enzo Cocca \email{enzo.ccc@gmail.com} (\href{https://orcid.org/0000-0002-8096-2810}{ORCID})
}

}
\keyword{internal}
