% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.sef_fit}
\alias{summary.sef_fit}
\title{Summarise a fitted SEF model}
\usage{
\method{summary}{sef_fit}(object, ...)
}
\arguments{
\item{object}{A \code{sef_fit} object.}

\item{...}{Ignored.}
}
\value{
A named list.
}
\description{
Summarise a fitted SEF model
}
