% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsestr-package.R
\docType{package}
\name{palimpsestr-package}
\alias{palimpsestr}
\alias{palimpsestr-package}
\title{palimpsestr: Probabilistic Decomposition of Archaeological Palimpsests}
\description{
Experimental tools for archaeological palimpsest analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/enzococca/palimpsestr}
  \item Report bugs at \url{https://github.com/enzococca/palimpsestr/issues}
}

}
\author{
\strong{Maintainer}: Enzo Cocca \email{enzo.cocca@cnr.it} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
