\name{palimpsestr-package}
\alias{palimpsestr-package}
\alias{palimpsestr}
\docType{package}
\title{palimpsestr: Probabilistic Decomposition of Archaeological Palimpsests}
\description{
Experimental tools for archaeological palimpsest analysis.
}
\details{
The package implements a prototype framework based on the Stratigraphic
Entanglement Field (SEF), the Stratigraphic Entanglement Index (SEI),
the Excavation Stratigraphic Energy (ESE), and the Palimpsest Dissolution
Index (PDI).
}
