\name{plot_energy}
\alias{plot_energy}
\title{Plot local Excavation Stratigraphic Energy}
\usage{plot_energy(object)}
\arguments{\item{object}{A \code{sef_fit} object.}}
\value{Invisibly returns the input object.}
\description{Creates a quick spatial diagnostic plot of local Excavation Stratigraphic Energy.}
