\name{plot_sei_profile}
\alias{plot_sei_profile}
\title{Plot ordered SEI profile}
\usage{plot_sei_profile(object)}
\arguments{\item{object}{A \code{sef_fit} object.}}
\value{Invisibly returns the input object.}
\description{Plot local SEI values after sorting observations by decreasing entanglement.}
