% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sei_profile}
\alias{plot_sei_profile}
\title{Plot ordered SEI profile (base R)}
\usage{
plot_sei_profile(object)
}
\arguments{
\item{object}{A \code{sef_fit} object.}
}
\value{
Invisibly returns the object.
}
\description{
Plot ordered SEI profile (base R)
}
\examples{
x <- archaeo_sim(n = 60, k = 2, seed = 1)
fit <- fit_sef(x, k = 2)
plot_sei_profile(fit)
}
\seealso{
\code{\link{sei_matrix}}, \code{\link{local_sei}}

Other plotting: 
\code{\link{as_plotly}()},
\code{\link{gg_compare_k}()},
\code{\link{gg_confusion}()},
\code{\link{gg_convergence}()},
\code{\link{gg_energy}()},
\code{\link{gg_entropy}()},
\code{\link{gg_intrusions}()},
\code{\link{gg_map}()},
\code{\link{gg_phase_profile}()},
\code{\link{gg_phasefield}()},
\code{\link{plot_energy}()},
\code{\link{plot_entropy}()},
\code{\link{plot_phasefield}()}
}
\concept{plotting}
